SWC-6471: move dataset script element injection to the start of the entity page load instead of the end
