+++ src/main/java/org/sagebionetworks/web/client/view/EntityView.java
@@ -38,4 +38,8 @@ public interface EntityView extends IsWidget {
   void setEntityPageTopVisible(boolean isVisible);
 
   void setOpenTeamInvitesVisible(boolean isVisible);
+
+  void injectDatasetJsonLd(String elementContent);
+
+  void removeDatasetJsonLdElement();
 }
