+++ src/main/java/org/sagebionetworks/web/client/widget/table/v2/TableEntityWidgetViewImpl.java
@@ -1,9 +1,5 @@
 package org.sagebionetworks.web.client.widget.table.v2;
 
-import com.google.gwt.dom.client.Document;
-import com.google.gwt.dom.client.Element;
-import com.google.gwt.dom.client.ScriptElement;
-import com.google.gwt.event.logical.shared.AttachEvent;
 import com.google.gwt.uibinder.client.UiBinder;
 import com.google.gwt.uibinder.client.UiField;
 import com.google.gwt.user.client.ui.Button;
@@ -86,7 +82,6 @@ public class TableEntityWidgetViewImpl
   SubmissionViewScopeWidget submissionViewScopeWidget;
   TableEntityWidgetView.Presenter presenter;
   SynapseReactClientFullContextPropsProvider propsProvider;
-  ScriptElement datasetScriptElement;
 
   @Inject
   public TableEntityWidgetViewImpl(
@@ -110,16 +105,6 @@ public class TableEntityWidgetViewImpl
     );
     scopeCollapseCloseButton.addClickHandler(event ->
       this.presenter.toggleScopeCollapse()
-    );
-    addAttachHandler(
-      new AttachEvent.Handler() {
-        @Override
-        public void onAttachOrDetach(AttachEvent event) {
-          if (!event.isAttached()) {
-            removeDatasetJsonLdElement();
-          }
-        }
-      }
     );
   }
 
@@ -249,24 +234,5 @@ public class TableEntityWidgetViewImpl
   @Override
   public void setQueryWrapperPlotNavVisible(boolean visible) {
     plotNavContainer.setVisible(visible);
-  }
-
-  @Override
-  public void removeDatasetJsonLdElement() {
-    if (datasetScriptElement != null) {
-      Element head = Document.get().getElementsByTagName("head").getItem(0);
-      head.removeChild(datasetScriptElement);
-      datasetScriptElement = null;
-    }
-  }
-
-  @Override
-  public void injectDatasetJsonLd(String elementContent) {
-    removeDatasetJsonLdElement();
-    Element head = Document.get().getElementsByTagName("head").getItem(0);
-    datasetScriptElement = Document.get().createScriptElement();
-    datasetScriptElement.setType("application/ld+json");
-    datasetScriptElement.setInnerText(elementContent);
-    head.appendChild(datasetScriptElement);
   }
 }
